cart checkout component added

+++ src/app/components/cart/cart.ts
@@ -1,60 +1,74 @@
-import { ChangeDetectorRef, Component, inject, model, OnDestroy, OnInit, Signal, signal } from '@angular/core';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { InputNumberModule } from 'primeng/inputnumber';
-import { CartService } from '../../shared/services/cart';
-import { Subscription } from 'rxjs';
-import { IProduct } from '../../shared/interfaces/product.interface';
 import { CurrencyPipe } from '@angular/common';
-import { ICartItem, IGroupedCartItems } from '../../shared/interfaces/cart.interface';
+import { ChangeDetectorRef, Component, inject, OnInit } from '@angular/core';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
+import { InputNumberModule } from 'primeng/inputnumber';
+import { Subscription } from 'rxjs';
+import { ICartItem } from '../../shared/interfaces/cart.interface';
+import { CartService } from '../../shared/services/cart';
 
 @Component({
   selector: 'app-cart',
-  imports: [InputNumberModule, FormsModule, ReactiveFormsModule, CurrencyPipe,RouterModule],
+  imports: [InputNumberModule, FormsModule, ReactiveFormsModule, CurrencyPipe, RouterModule],
   templateUrl: './cart.html',
   styleUrl: './cart.scss'
 })
 export class Cart implements OnInit {
   cartService = inject(CartService)
   cartSubscription: Subscription = new Subscription()
-  groupedCartItemsArray = signal([] as IGroupedCartItems[])
+  groupedCartItemsArray = this.cartService.getCartFilterList()
   $groupIndex: number = 0;
   cdr = inject(ChangeDetectorRef)
   calculatedAmountSummary = this.cartService.calculatedAmount
 
   ngOnInit(): void {
-    this.groupedCartItemsArray.set(this.cartService.getCartFilterList())
-    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray())
+    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray)
   }
 
 
   selectAllItems(event: Event) {
     const checked = (event.target as HTMLInputElement).checked;
-    this.cartService.selectAllItems(this.groupedCartItemsArray(), checked)
-    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray())
+    this.cartService.selectAllItems(this.groupedCartItemsArray, checked)
+    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray)
 
   }
 
   deleteSubItem() {
-    this.groupedCartItemsArray().splice(this.$groupIndex, 1)
-    this.cartService.deleteProductIntoCart()
-    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray())
+    this.groupedCartItemsArray.splice(this.$groupIndex, 1)
+    this.cartService.updateGroupCartList(this.groupedCartItemsArray)
+    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray)
+    this.cartService.cartAction$.next('delete')
   }
 
   deleteAllItems() {
-    this.groupedCartItemsArray.set([])
-    this.cartService.deleteProductIntoCart()
-    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray())
+    if (this.groupedCartItemsArray.length) {
+      this.cartService.deleteProductIntoCart()
+      this.getCartGroupList()
+      this.cartService.getTotalAmountSummary(this.groupedCartItemsArray)
+      this.cartService.cartAction$.next('delete')
+    }
   }
 
   quantityChange(item: ICartItem) {
     this.cartService.calculatePrice(item)
-    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray())
+    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray)
   }
 
   checkboxChange() {
-    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray())
+    this.cartService.updateGroupCartList(this.groupedCartItemsArray)
+    this.cartService.getTotalAmountSummary(this.groupedCartItemsArray)
+  }
 
+  get isCheckoutEnabled() {
+    return !this.groupedCartItemsArray.some(x => x.isSelected)
+  }
+
+  getCartGroupList() {
+    this.groupedCartItemsArray = this.cartService.getCartFilterList()
+  }
+
+  get selectedItemCount() {
+    return this.groupedCartItemsArray.filter(x => x.isSelected).length
   }
 
 }
